Rename AI packages
