+++ src/main/java/org/openRealmOfStars/game/GameKeyAdapter.java
@@ -22,7 +22,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.KeyEvent;
 import java.awt.image.BufferedImage;
 
-import org.openRealmOfStars.AI.Mission.MissionList;
+import org.openRealmOfStars.ai.mission.MissionList;
 import org.openRealmOfStars.ambient.BridgeCommandType;
 import org.openRealmOfStars.audio.music.MusicPlayer;
 import org.openRealmOfStars.audio.soundeffect.SoundPlayer;
