+++ src/main/java/org/openRealmOfStars/player/fleet/Fleet.java
@@ -26,8 +26,8 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.Optional;
 
-import org.openRealmOfStars.AI.Mission.Mission;
-import org.openRealmOfStars.AI.PathFinding.AStarSearch;
+import org.openRealmOfStars.ai.mission.Mission;
+import org.openRealmOfStars.ai.pathfinding.AStarSearch;
 import org.openRealmOfStars.player.PlayerInfo;
 import org.openRealmOfStars.player.SpaceRace.SpaceRace;
 import org.openRealmOfStars.player.leader.Job;
