+++ src/test/java/org/openRealmOfStars/ai/mission/MissionListTest.java
@@ -1,4 +1,4 @@
-package org.openRealmOfStars.AI.Mission;
+package org.openRealmOfStars.ai.mission;
 /*
  * Open Realm of Stars game project
  * Copyright (C) 2017-2019 Tuomo Untinen
