+++ src/test/java/org/openRealmOfStars/player/fleet/FleetTest.java
@@ -23,7 +23,7 @@ import static org.junit.Assert.*;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.mockito.Mockito;
-import org.openRealmOfStars.AI.PathFinding.AStarSearch;
+import org.openRealmOfStars.ai.pathfinding.AStarSearch;
 import org.openRealmOfStars.player.PlayerInfo;
 import org.openRealmOfStars.player.SpaceRace.SpaceRace;
 import org.openRealmOfStars.player.ship.Ship;
