+++ src/test/java/org/openRealmOfStars/utilities/repository/MissionRepositoryTest.java
@@ -21,8 +21,8 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.openRealmOfStars.AI.Mission.MissionPhase;
-import org.openRealmOfStars.AI.Mission.MissionType;
+import org.openRealmOfStars.ai.mission.MissionPhase;
+import org.openRealmOfStars.ai.mission.MissionType;
 
 /**
 * Test for MissionRepository
